+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.3.0] - 2021-04-30
+### Added
+- New Task `library-imports` (#12)
+  - Scans the source files (files under `lib/src`) and test files
+  - If any import in these files references a dart file that is a top level
+  library, the task rejects
+  - Top-Level libraries are dart files under `lib`, except those placed in
+  `lib/src`
 ### Changed
 - Change Mocking framework to mocktail